app/pipeline/requirements.txt
Flask==3.0.0
Flask_Cors==3.0.10
pika==1.3.2
redis==5.0.1
scipy==1.11.3
opencv-python-headless==4.8.1.78
uWSGI==2.0.23
python-decouple==3.8
memory-profiler==0.61.0
requests==2.31.0
boto3==1.26.118